Roll code related to HierarchyView into one file
